xpp: start moving halide/simd declarations
