+++ include/xpp/copy.h
@@ -25,9 +25,6 @@ XPP_EXPORT XppStatus Xpp_CopyFromRetina_c(uint8_t* pDstData, int nDstStep, int n
 XPP_EXPORT XppStatus Xpp_Move_c(uint8_t* pData, int nStep, int nXDst, int nYDst, int nWidth, int nHeight, int nXSrc,
 				int nYSrc);
 
-XPP_EXPORT XppStatus Xpp_Copy_halide(uint8_t* pDstData, int nDstStep, int nXDst, int nYDst, int nWidth, int nHeight,
-				     uint8_t* pSrcData, int nSrcStep, int nXSrc, int nYSrc);
-
 #ifdef __cplusplus
 }
 #endif
